+++ src/plugins/relay/relay-config.h
@@ -61,6 +61,7 @@ extern struct t_config_option *relay_config_network_tls_priorities;
 extern struct t_config_option *relay_config_network_totp_secret;
 extern struct t_config_option *relay_config_network_totp_window;
 extern struct t_config_option *relay_config_network_websocket_allowed_origins;
+extern struct t_config_option *relay_config_network_websocket_permessage_deflate;
 
 extern struct t_config_option *relay_config_irc_backlog_max_minutes;
 extern struct t_config_option *relay_config_irc_backlog_max_number;