+++ CHANGELOG.md
@@ -4,8 +4,9 @@
 
 ### Added
 
-- relay: add infos "relay_api_version" and "relay_api_version_number"
+- relay: add option relay.network.websocket_permessage_deflate
 - relay: add option relay.look.raw_messages_max_length ([#2122](https://github.com/weechat/weechat/issues/2122))
+- relay: add infos "relay_api_version" and "relay_api_version_number"
 
 ### Fixed